Refactor: Extract TransactionRow and Pagination components
- Extract TransactionRow component from transactions page
- Extract Pagination component for reusable pagination UI
- Move formatting functions (formatDate, formatAmount, getTransactionTypeLabel) to TransactionRow
- Simplify transactions page by using extracted components
- Improve code reusability and maintainability
